- Added .next into .gitignore and remove .next directory from temp/web folder - Move theme folder into root packages folder - Moved all configs under root/packages/config
