+++ .gitignore
@@ -10,3 +10,4 @@ reports
 node_modules/
 package-lock.json
 .DS_Store
+.next
