+++ templates/base/package.json
@@ -34,7 +34,8 @@
     "react": "*",
     "react-dom": "*",
     "styled-jsx": "4.0.1",
-    "styled-components": "^5.3.1"
+    "styled-components": "^5.3.1",
+    "themes": "*"
   },
   "devDependencies": {
     "@next/bundle-analyzer": "11.1.2",
