+++ templates/base/src/pages/_app.js
@@ -5,7 +5,7 @@ import GraphQLClient from '../graphQL/ApolloClient'
 import GlobalContextProvider from '../stores/global/ContextProvider';
 
 import { ThemeProvider } from 'styled-components';
-import { theme } from '../styles/themes/theme';
+import { theme } from 'themes';
 
 // Will be called once for every metric that has to be reported.
 export function reportWebVitals(metric) {
