--- templates/common/essentials/config/bundlesize.json
@@ -1,9 +1,0 @@
-{
-  "files": [
-    {
-      "path": "./build/**/chunks/*.js",
-      "maxSize": "50 kB",
-      "compression": "gzip"
-    }
-  ]
-}
