--- templates/common/essentials/config/endpoints.js
@@ -1,16 +1,0 @@
-/** list of apis to be called */
-export const endpoints = {
-  content: {
-    home: 'home/1'
-  },
-  apiGee: {
-    todo: 'todos/1'
-  }
-};
-
-/** various domains entry
- * 1. contentDomain - for content type like labels, api call to graphql, etc
- * 2. dataDomain - for data type, like api call for user data
- */
-export const contentDomain = 'https://jsonplaceholder.typicode.com/';
-export const apiGeeDomain = 'https://jsonplaceholder.typicode.com/';
