--- templates/common/essentials/config/jest/jest.config.js
@@ -1,40 +1,0 @@
-module.exports = {
-    rootDir: '../../',
-    moduleDirectories: ['node_modules'],
-    testEnvironment: 'jsdom',
-    setupFiles: [
-      "<rootDir>/config/jest/jest.setup.js"
-    ],
-    moduleNameMapper: {
-      "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
-      "\\.(css|scss)$": "<rootDir>/__mocks__/styleMock.js"
-    },
-    testPathIgnorePatterns: ['<rootDir>/dist', '<rootDir>/out', '<rootDir>/build'],
-    clearMocks: true,
-    collectCoverage: true,
-    coverageDirectory: '<rootDir>/reports/coverage',
-    collectCoverageFrom: [
-      'src/**/*.js',
-      'src/**/*.jsx',
-      'store/**/*.js', 
-      '!**/*.mock.js',
-      '!**/*.story.js',
-      '!**/*.styles.js',
-      '!**styles/**/*.js',
-      '!**/node_modules/**'
-    ],
-    coveragePathIgnorePatterns: ['/node_modules/'],
-    coverageReporters: ['lcov', 'json', 'text-summary'],
-    coverageThreshold: {
-      global: {
-        "branches": 30,
-        "functions": 30,
-        "lines": 30,
-        "statements": 30
-      }
-    },
-    transform: {
-      '\\.(js|jsx)?$': 'babel-jest'
-    }
-  };
-
