--- templates/common/essentials/config/jest/jest.setup.js
@@ -1,5 +1,0 @@
-const mockFetchPromise = Promise.resolve({
-    json: () => Promise.resolve({}),
-  });
-  global.fetch = jest.fn(() => mockFetchPromise);
-
