--- templates/common/essentials/config/performance/sitespeed.budget.config.json
@@ -1,11 +1,0 @@
-{
-  "budget": {
-    "lighthouse": {
-      "performance": 80,
-      "accessibility": 90,
-      "best-practices": 90,
-      "seo": 90,
-      "pwa": 10
-    }
-  }
-}
