--- templates/common/essentials/config/performance/urls.txt
@@ -1,1 +1,0 @@
-https://pagesource.github.io/universal-react-docs/
