+++ templates/common/packages/config/jest/jest.setup.js
@@ -1,0 +1,5 @@
+const mockFetchPromise = Promise.resolve({
+    json: () => Promise.resolve({}),
+  });
+  global.fetch = jest.fn(() => mockFetchPromise);
+
