+++ templates/common/packages/config/performance/sitespeed.budget.config.json
@@ -1,0 +1,11 @@
+{
+  "budget": {
+    "lighthouse": {
+      "performance": 80,
+      "accessibility": 90,
+      "best-practices": 90,
+      "seo": 90,
+      "pwa": 10
+    }
+  }
+}
