+++ templates/common/packages/config/performance/urls.txt
@@ -1,0 +1,1 @@
+https://pagesource.github.io/universal-react-docs/
