+++ templates/common/packages/themes/theme.js
@@ -1,0 +1,927 @@
+export const theme = {
+    "meta": {
+        "remFactor": 16
+    },
+    "measure": {
+        "narrow": "20em",
+        "default": "28em",
+        "wide": "34em"
+    },
+    "lineHeights": {
+        "solid": 1,
+        "body": 1.5,
+        "heading": 1.25
+    },
+    "fonts": [
+        {
+            "value": "\"FuturaNextW05-Bold\",\"Helvetica\",\"Arial\",sans-serif",
+            "name": "Font Family - 0",
+            "id": "Yl2JxoCsmL"
+        },
+        {
+            "value": "\"FuturaNextW05-Book\",\"Helvetica\",\"Arial\",sans-serif",
+            "name": "Font Family - 1",
+            "id": "DAZWpdGkYs"
+        },
+        {
+            "value": "\"FuturaNextW05-Book\",\"Helvetica\",\"Arial\",sans-serif,sans-serif,Arial",
+            "name": "Font Family - 2",
+            "id": "u0wIDC6dzr"
+        },
+        {
+            "value": "\"FuturaNextW05-Book\",BlinkMacSystemFont,-apple-system,\"Segoe UI\",\"Roboto\",\"Oxygen\",\"Ubuntu\",\"Cantarell\",\"Fira Sans\",\"Droid Sans\",\"Helvetica Neue\",\"Helvetica\",\"Arial\",sans-serif",
+            "name": "Font Family - 3",
+            "id": "NBg2sg0N8y"
+        },
+        {
+            "value": "\"FuturaNextW05-DemiBold\",\"Helvetica\",\"Arial\",sans-serif",
+            "name": "Font Family - 4",
+            "id": "fJD0aB2Dv4"
+        },
+        {
+            "value": "\"FuturaNextW05-ExtraBold\",\"Helvetica\",\"Arial\",sans-serif",
+            "name": "Font Family - 5",
+            "id": "9ipXGxqFF4"
+        },
+        {
+            "value": "\"FuturaNextW05-Light\",\"Helvetica\",\"Arial\",sans-serif",
+            "name": "Font Family - 6",
+            "id": "MN3Ibd82v3"
+        },
+        {
+            "value": "\"FuturaNextW05-Medium\",\"Helvetica\",\"Arial\",sans-serif",
+            "name": "Font Family - 7",
+            "id": "Plyld5mFaT"
+        },
+        {
+            "value": "\"minion-pro\",serif",
+            "name": "Font Family - 8",
+            "id": "U88eU3EMyd"
+        },
+        {
+            "value": "'Font Awesome 5 Free'",
+            "name": "Font Family - 9",
+            "id": "zT3BSVxLhO"
+        },
+        {
+            "value": "monospace",
+            "name": "Font Family - 10",
+            "id": "3n0i3oUdEG"
+        },
+        {
+            "value": "source-han-sans-cjk-ja, sans-serif",
+            "name": "Font Family - 11",
+            "id": "PYxSPlCBiB"
+        },
+        {
+            "value": "source-han-sans-simplified-c, sans-serif",
+            "name": "Font Family - 12",
+            "id": "fuMRvKjkA6"
+        },
+        {
+            "value": "source-han-serif-japanese, serif",
+            "name": "Font Family - 13",
+            "id": "pYGQWxC3jD"
+        },
+        {
+            "value": "source-han-serif-tc, serif",
+            "name": "Font Family - 14",
+            "id": "O7zm29zclk"
+        }
+    ],
+    "colors": {
+        "white": [
+            "white",
+            "#fff",
+            "#fffef5",
+            "#f5fffd",
+            "#f5fbff",
+            "#fafafa",
+            "#f9f9f9",
+            "#fff5f5",
+            "#f5f5f5",
+            "#F4F4F4",
+            "#f2f2f2",
+            "#f1f1f1",
+            "#eee",
+            "#e8e8e8",
+            "#e6e6e6",
+            "#dedede",
+            "#dbdbdb"
+        ],
+        "black": [
+            "#242424",
+            "#1c1c1c",
+            "#171403",
+            "#121212",
+            "#000403",
+            "black",
+            "#000",
+            "#000000"
+        ],
+        "red": [
+            "#ff414d",
+            "#fe414d",
+            "#fe3441",
+            "#fe2835",
+            "#fe0e1d",
+            "red",
+            "#d00a17"
+        ],
+        "gray": [
+            "#b5b5b5",
+            "#b4b4b4",
+            "#acacac",
+            "#9B9B9B",
+            "#666",
+            "#565656",
+            "#505050",
+            "#4d4d4d",
+            "#4A4A4A",
+            "#363636",
+            "#333",
+            "#2f2f2f",
+            "#2a2a2a",
+            "#292929"
+        ],
+        "cyan": [
+            "#00e6c3",
+            "#00e6c2",
+            "#00d9b8",
+            "#07d6b6",
+            "#00cdad",
+            "#00b398",
+            "#22A0FF",
+            "#079fff",
+            "#0099f9",
+            "#0091ed",
+            "#0081d3",
+            "#074e7b"
+        ],
+        "yellow": [
+            "#ffe63b",
+            "#ffe42e",
+            "#ffe322",
+            "#ffdf08"
+        ],
+        "overlays": [
+            "rgba(255,255,255,0.9)",
+            "rgba(255,255,255,0.7)",
+            "rgba(245,245,245,0.9)",
+            "rgba(245,245,245,0.7)",
+            "rgba(238,238,238,0.3)",
+            "rgba(54,54,54,0.9)",
+            "rgba(54,54,54,0.3)",
+            "rgba(54,54,54,0.7)",
+            "rgba(0,0,0,0.83)",
+            "rgba(0,0,0,0.1)",
+            "rgba(0,0,0,0.86)",
+            "rgba(0,0,0,0.4)",
+            "rgba(0,0,0,0.2)",
+            "rgba(0,0,0,0.7)",
+            "rgba(0,0,0,0.9)",
+            "rgba(0,0,0,0.3)",
+            "rgba(0,0,0,0)",
+            "rgba(0,0,0,0.05)",
+            "rgba(0,0,0,0.65)"
+        ],
+        "modes": {
+
+        }
+    },
+    "radii": [
+        "1.6px",
+        "2px",
+        "4px",
+        "6px",
+        "12px",
+        "16px",
+        "20px",
+        "22px",
+        "24px",
+        "25px",
+        "40px",
+        "50px",
+        "290486px"
+    ],
+    "space": [
+        "1px",
+        "2px",
+        "2.4px",
+        "3px",
+        "3.2px",
+        "4px",
+        "4.8px",
+        "5px",
+        "6px",
+        "7px",
+        "8px",
+        "10px",
+        "12px",
+        "14px",
+        "14.4px",
+        "15px",
+        "16px",
+        "20px",
+        "22px",
+        "22.4px",
+        "24px",
+        "26px",
+        "30px",
+        "32px",
+        "40px",
+        "48px",
+        "50px",
+        "56px",
+        "60px",
+        "62px",
+        "64px",
+        "70px",
+        "80px",
+        "96px",
+        "100px",
+        "128px",
+        "140px",
+        "144px",
+        "180px",
+        "200px",
+        "288px"
+    ],
+    "fontSizes": [
+        "10px",
+        "11px",
+        "12px",
+        "13px",
+        "14px",
+        "15px",
+        "16px",
+        "17px",
+        "18px",
+        "20px",
+        "21px",
+        "21.33328px",
+        "22px",
+        "23px",
+        "24px",
+        "26px",
+        "28px",
+        "30px",
+        "32px",
+        "34px",
+        "35px",
+        "36px",
+        "38px",
+        "40px",
+        "42px",
+        "44px",
+        "48px",
+        "64px",
+        "65px",
+        "80px"
+    ],
+    "borderWidths": [
+        "1px",
+        "2px",
+        "3px",
+        "4px"
+    ],
+    "fontWeights": [
+        100,
+        200,
+        300,
+        400,
+        500,
+        600,
+        700,
+        900
+    ],
+    "gradients": [
+        "linear-gradient(141deg, #e6e6e6 0%, #fff 71%, #fff 100%)",
+        "linear-gradient(141deg, #000 0%, #000 71%, #0d0d0d 100%)",
+        "linear-gradient(141deg, #dfd8d9 0%, #f5f5f5 71%, #fff 100%)",
+        "linear-gradient(141deg, #1f191a 0%, #363636 71%, #46403f 100%)",
+        "linear-gradient(141deg, #ff0d45 0%, #fe414d 71%, #ff6b5a 100%)",
+        "linear-gradient(141deg, #00a4d3 0%, #079fff 71%, #2184ff 100%)",
+        "linear-gradient(141deg, #00b37a 0%, #00e6c3 71%, #01fbff 100%)",
+        "linear-gradient(141deg, #ffb608 0%, #ffe63b 71%, #f8ff55 100%)"
+    ],
+    "borderStyles": [
+        "solid"
+    ],
+    "boxShadows": [
+        {
+            "value": "none",
+            "name": "boxshadow-0"
+        },
+        {
+            "value": "0 2px 3px rgba(0,0,0,0.1),0 0 0 1px rgba(0,0,0,0.1)",
+            "name": "boxshadow-1"
+        },
+        {
+            "value": "0 2px 3px rgba(0,0,0,0.1),0 0 0 1px #fe414d",
+            "name": "boxshadow-2"
+        },
+        {
+            "value": "inset 0 1px 2px rgba(0,0,0,0.2),0 0 0 1px #fe414d",
+            "name": "boxshadow-3"
+        },
+        {
+            "value": "0 0 0 0.125em rgba(254,65,77,0.25)",
+            "name": "boxshadow-4"
+        },
+        {
+            "value": "0 0 0 0.125em rgba(255,255,255,0.25)",
+            "name": "boxshadow-5"
+        },
+        {
+            "value": "0 0 0 0.125em rgba(0,0,0,0.25)",
+            "name": "boxshadow-6"
+        },
+        {
+            "value": "0 0 0 0.125em rgba(245,245,245,0.25)",
+            "name": "boxshadow-7"
+        },
+        {
+            "value": "0 0 0 0.125em rgba(54,54,54,0.25)",
+            "name": "boxshadow-8"
+        },
+        {
+            "value": "0 0 0 0.125em rgba(7,159,255,0.25)",
+            "name": "boxshadow-9"
+        },
+        {
+            "value": "0 0 0 0.125em rgba(0,230,195,0.25)",
+            "name": "boxshadow-10"
+        },
+        {
+            "value": "0 0 0 0.125em rgba(255,230,59,0.25)",
+            "name": "boxshadow-11"
+        },
+        {
+            "value": "inset 0 1px 2px rgba(0,0,0,0.1)",
+            "name": "boxshadow-12"
+        },
+        {
+            "value": "0 0 0.5em rgba(255,255,255,0.25)",
+            "name": "boxshadow-13"
+        },
+        {
+            "value": "0 0 0.5em rgba(0,0,0,0.25)",
+            "name": "boxshadow-14"
+        },
+        {
+            "value": "0 0 0.5em rgba(245,245,245,0.25)",
+            "name": "boxshadow-15"
+        },
+        {
+            "value": "0 0 0.5em rgba(54,54,54,0.25)",
+            "name": "boxshadow-16"
+        },
+        {
+            "value": "0 0 0.5em rgba(254,65,77,0.25)",
+            "name": "boxshadow-17"
+        },
+        {
+            "value": "0 0 0.5em rgba(7,159,255,0.25)",
+            "name": "boxshadow-18"
+        },
+        {
+            "value": "0 0 0.5em rgba(0,230,195,0.25)",
+            "name": "boxshadow-19"
+        },
+        {
+            "value": "0 0 0.5em rgba(255,230,59,0.25)",
+            "name": "boxshadow-20"
+        },
+        {
+            "value": "0 1px 2px rgba(0,0,0,0.1)",
+            "name": "boxshadow-21"
+        },
+        {
+            "value": "0 2px 0 0 #f5f5f5",
+            "name": "boxshadow-22"
+        },
+        {
+            "value": "0 -2px 0 0 #f5f5f5",
+            "name": "boxshadow-23"
+        },
+        {
+            "value": "0 8px 16px rgba(0,0,0,0.1)",
+            "name": "boxshadow-24"
+        },
+        {
+            "value": "0 -2px 3px rgba(0,0,0,0.1)",
+            "name": "boxshadow-25"
+        },
+        {
+            "value": "0 -8px 8px rgba(0,0,0,0.1)",
+            "name": "boxshadow-26"
+        },
+        {
+            "value": "0 8px 8px rgba(0,0,0,0.1)",
+            "name": "boxshadow-27"
+        },
+        {
+            "value": "0 8px 8px rgba(0,0,0,0.1),0 0 0 1px rgba(0,0,0,0.1)",
+            "name": "boxshadow-28"
+        },
+        {
+            "value": "inset 0 1px 2px rgba(0,0,0,0.2)",
+            "name": "boxshadow-29"
+        },
+        {
+            "value": "0px 0px 3px #999",
+            "name": "boxshadow-30"
+        }
+    ],
+    "textShadows": [
+        {
+            "value": "0 0 0 #000",
+            "name": "textshadows-0"
+        }
+    ],
+    "borderColors": [
+        "yellow",
+        "rgba(255,255,255,0.5)",
+        "#cfcfcf",
+        "#d5d5d5",
+        "#eee",
+        "#ffe322",
+        "#00cdad",
+        "#0091ed",
+        "#fe2835",
+        "#292929",
+        "#e8e8e8",
+        "#f2f2f2",
+        "#ffe63b",
+        "rgba(0,0,0,0.7)",
+        "#00e6c3",
+        "#079fff",
+        "#fe414d",
+        "#363636",
+        "#f5f5f5",
+        "#000",
+        "#fff",
+        "transparent",
+        "#b4b4b4",
+        "#b5b5b5",
+        "#dbdbdb"
+    ],
+    "transitions": [
+        "all 1s cubic-bezier(0.32, 0.01, 0, 1)",
+        "all 0.5s cubic-bezier(0.39, 0.575, 0.565, 1)",
+        "all 0.3s cubic-bezier(0.32, 0.01, 0, 1)",
+        "background-size 0.25s 0s cubic-bezier(0.32, 0.01, 0, 1)",
+        "all .25s linear",
+        "transform 200ms linear",
+        "all 0.35s cubic-bezier(0.32, 0.01, 0, 1)",
+        "transform 0.3s",
+        "transform 0.7s cubic-bezier(0.32, 0.01, 0, 1)",
+        "top 0.7s cubic-bezier(0.32, 0.01, 0, 1)",
+        "color 0.45s cubic-bezier(0.32, 0.01, 0, 1)",
+        "none",
+        "right .8s",
+        "height 0.5s 0s cubic-bezier(0.32, 0.01, 0, 1)",
+        "opacity 0.25s linear",
+        "all 0.25s 0s cubic-bezier(0.32, 0.01, 0, 1)",
+        "all 1s ease-in-out",
+        "opacity 0.6s ease-in-out",
+        "ease-in-out 0.2s",
+        "all 0.45s cubic-bezier(0.32, 0.01, 0, 1)",
+        "all 0.2s",
+        "right 0.45s cubic-bezier(0.32, 0.01, 0, 1)",
+        "all 0.5s cubic-bezier(0.32, 0.01, 0, 1)",
+        "transform 0.3s cubic-bezier(0.32, 0.01, 0, 1)",
+        "opacity .8s",
+        "background-color .2s ease-out, border-color .2s ease-out, color .2s ease-out",
+        "all 0.3s ease-in-out",
+        "all .6s ease-in-out",
+        "background-color 0.3s cubic-bezier(0.32, 0.01, 0, 1)"
+    ],
+    "animations": [
+        "fa-spin 2s infinite linear",
+        "fa-spin 1s infinite steps(8)",
+        "spinAround 500ms infinite linear",
+        "fadein 2s",
+        "fadeout 2s"
+    ],
+    "transitionDelay": [
+        "0ms",
+        "300ms",
+        "200ms"
+    ],
+    "transitionProperty": [
+        "background-color",
+        "opacity",
+        "transform",
+        "visibility",
+        "all",
+        "background-size",
+        "top",
+        "color",
+        "none",
+        "right",
+        "height",
+        "border-color"
+    ],
+    "transitionDuration": [
+        "0ms",
+        "86ms",
+        "200ms",
+        "250ms",
+        "300ms",
+        "350ms",
+        "450ms",
+        "500ms",
+        "600ms",
+        "700ms",
+        "800ms",
+        "1000ms"
+    ],
+    "transitionTimingFunc": [
+        "ease-out",
+        "cubic-bezier(0.32, 0.01, 0, 1)",
+        "cubic-bezier(0.39, 0.575, 0.565, 1)",
+        "linear",
+        "ease",
+        "ease-in-out"
+    ],
+    "containerWidth": 1440,
+    "breakpoints": [
+        {
+            "value": 600,
+            "name": "breakpoint-0"
+        },
+        {
+            "value": 1024,
+            "name": "breakpoint-1"
+        }
+    ],
+    "styles": {
+
+    },
+    "flutterThemeData": {
+        "accentColor": {
+            "type": "Color",
+            "value": "#4A4A4A",
+            "locked": false
+        },
+        "backgroundColor": {
+            "type": "Color",
+            "value": "#242424",
+            "locked": false
+        },
+        "bottomAppBarColor": {
+            "type": "Color",
+            "value": "#dfded5",
+            "locked": false
+        },
+        "buttonColor": {
+            "type": "Color",
+            "value": "#FFFEF5",
+            "locked": false
+        },
+        "canvasColor": {
+            "type": "Color",
+            "value": "#242424",
+            "locked": false
+        },
+        "cardColor": {
+            "type": "Color",
+            "value": "#565656",
+            "locked": false
+        },
+        "cursorColor": {
+            "type": "Color",
+            "value": "#171403",
+            "locked": false
+        },
+        "dialogBackgroundColor": {
+            "type": "Color",
+            "value": "#565656",
+            "locked": false
+        },
+        "disabledColor": {
+            "type": "Color",
+            "value": "#ffffff",
+            "locked": false
+        },
+        "dividerColor": {
+            "type": "Color",
+            "value": "#F4F4F4",
+            "locked": false
+        },
+        "errorColor": {
+            "type": "Color",
+            "value": "#FE2835",
+            "locked": false
+        },
+        "focusColor": {
+            "type": "Color",
+            "value": "#FFFEF5",
+            "locked": false
+        },
+        "highlightColor": {
+            "type": "Color",
+            "value": "#ffffff",
+            "locked": false
+        },
+        "hoverColor": {
+            "type": "Color",
+            "value": "#ffffff",
+            "locked": false
+        },
+        "indicatorColor": {
+            "type": "Color",
+            "value": "#4A4A4A",
+            "locked": false
+        },
+        "primaryColor": {
+            "type": "Color",
+            "value": "#FFFEF5",
+            "locked": false
+        },
+        "primaryColorDark": {
+            "type": "Color",
+            "value": "#dfded5",
+            "locked": false
+        },
+        "primaryColorLight": {
+            "type": "Color",
+            "value": "#ffffff",
+            "locked": false
+        },
+        "scaffoldBackgroundColor": {
+            "type": "Color",
+            "value": "#242424",
+            "locked": false
+        },
+        "secondaryHeaderColor": {
+            "type": "Color",
+            "value": "#F4F4F4",
+            "locked": false
+        },
+        "selectedRowColor": {
+            "type": "Color",
+            "value": "#4A4A4A",
+            "locked": false
+        },
+        "splashColor": {
+            "type": "Color",
+            "value": "#ffffff",
+            "locked": false
+        },
+        "textSelectionColor": {
+            "type": "Color",
+            "value": "#171403",
+            "locked": false
+        },
+        "textSelectionHandleColor": {
+            "type": "Color",
+            "value": "#171403",
+            "locked": false
+        },
+        "toggleableActiveColor": {
+            "type": "Color",
+            "value": "#4A4A4A",
+            "locked": false
+        },
+        "textTheme": {
+            "headline1": {
+                "color": {
+                    "value": "#171403"
+                },
+                "fontSize": {
+                    "value": ""
+                },
+                "fontStyle": {
+                    "value": ""
+                },
+                "fontFamily": {
+                    "value": ""
+                },
+                "fontWeight": {
+                    "value": ""
+                },
+                "locked": false
+            },
+            "headline2": {
+                "color": {
+                    "value": "#171403"
+                },
+                "fontSize": {
+                    "value": ""
+                },
+                "fontStyle": {
+                    "value": ""
+                },
+                "fontFamily": {
+                    "value": ""
+                },
+                "fontWeight": {
+                    "value": ""
+                },
+                "locked": false
+            },
+            "headline3": {
+                "color": {
+                    "value": "#171403"
+                },
+                "fontSize": {
+                    "value": ""
+                },
+                "fontStyle": {
+                    "value": ""
+                },
+                "fontFamily": {
+                    "value": ""
+                },
+                "fontWeight": {
+                    "value": ""
+                },
+                "locked": false
+            },
+            "headline4": {
+                "color": {
+                    "value": "#171403"
+                },
+                "fontSize": {
+                    "value": ""
+                },
+                "fontStyle": {
+                    "value": ""
+                },
+                "fontFamily": {
+                    "value": ""
+                },
+                "fontWeight": {
+                    "value": ""
+                },
+                "locked": false
+            },
+            "headline5": {
+                "color": {
+                    "value": "#171403"
+                },
+                "fontSize": {
+                    "value": ""
+                },
+                "fontStyle": {
+                    "value": ""
+                },
+                "fontFamily": {
+                    "value": ""
+                },
+                "fontWeight": {
+                    "value": ""
+                },
+                "locked": false
+            },
+            "headline6": {
+                "color": {
+                    "value": "#171403"
+                },
+                "fontSize": {
+                    "value": ""
+                },
+                "fontStyle": {
+                    "value": ""
+                },
+                "fontFamily": {
+                    "value": ""
+                },
+                "fontWeight": {
+                    "value": ""
+                },
+                "locked": false
+            },
+            "subtitle1": {
+                "color": {
+                    "value": "#171403"
+                },
+                "fontSize": {
+                    "value": ""
+                },
+                "fontStyle": {
+                    "value": ""
+                },
+                "fontFamily": {
+                    "value": ""
+                },
+                "fontWeight": {
+                    "value": ""
+                },
+                "locked": false
+            },
+            "subtitle2": {
+                "color": {
+                    "value": "#171403"
+                },
+                "fontSize": {
+                    "value": ""
+                },
+                "fontStyle": {
+                    "value": ""
+                },
+                "fontFamily": {
+                    "value": ""
+                },
+                "fontWeight": {
+                    "value": ""
+                },
+                "locked": false
+            },
+            "body1": {
+                "color": {
+                    "value": "#F4F4F4"
+                },
+                "fontSize": {
+                    "value": ""
+                },
+                "fontStyle": {
+                    "value": ""
+                },
+                "fontFamily": {
+                    "value": ""
+                },
+                "fontWeight": {
+                    "value": ""
+                },
+                "locked": false
+            },
+            "body2": {
+                "color": {
+                    "value": "#F4F4F4"
+                },
+                "fontSize": {
+                    "value": ""
+                },
+                "fontStyle": {
+                    "value": ""
+                },
+                "fontFamily": {
+                    "value": ""
+                },
+                "fontWeight": {
+                    "value": ""
+                },
+                "locked": false
+            },
+            "button": {
+                "color": {
+                    "value": "#171403"
+                },
+                "fontSize": {
+                    "value": ""
+                },
+                "fontStyle": {
+                    "value": ""
+                },
+                "fontFamily": {
+                    "value": ""
+                },
+                "fontWeight": {
+                    "value": ""
+                },
+                "locked": false
+            },
+            "caption": {
+                "color": {
+                    "value": "#F4F4F4"
+                },
+                "fontSize": {
+                    "value": ""
+                },
+                "fontStyle": {
+                    "value": ""
+                },
+                "fontFamily": {
+                    "value": ""
+                },
+                "fontWeight": {
+                    "value": ""
+                },
+                "locked": false
+            },
+            "overline": {
+                "color": {
+                    "value": "#F4F4F4"
+                },
+                "fontSize": {
+                    "value": ""
+                },
+                "fontStyle": {
+                    "value": ""
+                },
+                "fontFamily": {
+                    "value": ""
+                },
+                "fontWeight": {
+                    "value": ""
+                },
+                "locked": false
+            }
+        }
+    },
+    "originURL": "google.com"
+}
